Move options to root package
Just like `RecheckImpl` and `RecheckOptions` reside in the same package `de.retest.recheck`, `RecheckWebImpl` and `RecheckWebOptions` should do the same.
